remove menu authorized to menu-model
