--- src/utils/FilterMenuData.js
@@ -1,37 +1,0 @@
-import Authorized from './Authorized';
-
-const { check } = Authorized;
-
-/**
- * get SubMenu or Item
- */
-const getSubMenu = item => {
-  // doc: add hideChildrenInMenu
-  if (item.children && !item.hideChildrenInMenu && item.children.some(child => child.name)) {
-    return {
-      ...item,
-      children: filterMenuData(item.children), // eslint-disable-line
-    };
-  }
-  return item;
-};
-
-/**
- * filter menuData
- */
-const filterMenuData = menuData => {
-  if (!menuData) {
-    return [];
-  }
-  return menuData
-    .filter(item => item.name && !item.hideInMenu)
-    .map(item => {
-      // make dom
-      const ItemDom = getSubMenu(item);
-      const data = check(item.authority, ItemDom);
-      return data;
-    })
-    .filter(item => item);
-};
-
-export default filterMenuData;
